switchctl: Use static inline func defs
These are required when compiling with -Os or other modes that
would instead call this function rather than actually inline it.

This is exposed due to not using the function as inline along with
how autotools ends up building the final binary. Adding static
to the definition ends up being more portable.