scripts: remove unnecessary scripts, rename tssupercapmon to tssilomon
The removed scripts were systemd scripts targetted at older kernels
and distributions in order to configure the ethernet switch in a
flexible way. Newer kernels have this switch directly supported and
are able to configure the interfaces in a standard Linux way.

The tssupercapmon script is named tssilomon everywhere else, this
is renamed for uniformity.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>